# backend/openapi.yaml
openapi: 3.0.3
info:
  title: Beat Portal API
  description: |
    FastAPI backend for Beat Portal - A DJ music library organizer with AI-powered metadata enhancement.
    
    ## Features
    - Offline-first music library management
    - Advanced metadata analysis (BPM, Key, Genre, Mood)
    - AI-powered metadata enhancement
    - Track organization and playlist management
  version: 1.0.0
  contact:
    name: Beat Portal Support
    email: support@beatportal.dev
servers:
  - url: http://localhost:8000/api/v1
    description: Local development server
  - url: http://localhost:8000
    description: Local base server
...
paths:
...
  /playlists:
    get:
      tags:
        - Playlists
      summary: List Playlists
      description: Get all playlists
      responses:
        '200':
          description: Playlists retrieved successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Playlist'

    post:
      tags:
        - Playlists
      summary: Create Playlist
      description: Create a new playlist
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreatePlaylistRequest'
      responses:
        '201':
          description: Playlist created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Playlist'
  /playlists/{playlist_id}:
    get:
      tags:
        - Playlists
      summary: Get Playlist
      description: Get playlist details including tracks
      parameters:
        - name: playlist_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Playlist retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PlaylistDetail'

    put:
      tags:
        - Playlists
      summary: Update Playlist
      description: Update playlist name or description
      parameters:
        - name: playlist_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdatePlaylistRequest'
      responses:
        '200':
          description: Playlist updated successfully

    delete:
      tags:
        - Playlists
      summary: Delete Playlist
      description: Delete a playlist
      parameters:
        - name: playlist_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Playlist deleted successfully
  /playlists/{playlist_id}/tracks:
    post:
      tags:
        - Playlists
      summary: Add Tracks to Playlist
      description: Add one or more tracks to a playlist
      parameters:
        - name: playlist_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AddTracksToPlaylistRequest'
      responses:
        '200':
          description: Tracks added successfully

    delete:
      tags:
        - Playlists
      summary: Remove Tracks from Playlist
      description: Remove tracks from a playlist
      parameters:
        - name: playlist_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RemoveTracksFromPlaylistRequest'
      responses:
        '200':
          description: Tracks removed successfully
  /playlists/{playlist_id}/export:
    post:
      tags:
        - Playlists
      summary: Export Playlist
      description: Export playlist in various formats
      parameters:
        - name: playlist_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ExportPlaylistRequest'
      responses:
        '200':
          description: Playlist exported successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExportPlaylistResponse'
...
components:
  schemas:
    Track:
      type: object
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
        artist:
          type: string
        album:
          type: string
        year:
          type: integer
        genre:
          type: string
        mood:
          type: string
        bpm:
          type: integer
        key:
          type: string
        duration_seconds:
          type: integer
          description: Duration in seconds
        file_path:
          type: string
        file_size_bytes:
          type: integer
          description: File size in bytes
        file_format:
          type: string
          example: mp3
        bitrate_bps:
          type: integer
          description: Bitrate in bits per second
        sample_rate_hz:
          type: integer
          description: Sample rate in Hertz (Hz)
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        last_played:
          type: string
          format: date-time
        play_count:
          type: integer
        metadata_complete:
          type: boolean
...
    Playlist:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        description:
          type: string
        track_count:
          type: integer
        total_duration_seconds:
          type: integer
          description: Total duration in seconds
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    PlaylistDetail:
      allOf:
        - $ref: '#/components/schemas/Playlist'
        - type: object
          properties:
            tracks:
              type: array
              items:
                $ref: '#/components/schemas/Track'
...
    CreatePlaylistRequest:
      type: object
      required:
        - name
      properties:
        name:
          type: string
          example: "Summer House Mix"
        description:
          type: string
          example: "High energy house tracks for summer sets"
    UpdatePlaylistRequest:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
    AddTracksToPlaylistRequest:
      type: object
      properties:
        track_ids:
          type: array
          items:
            type: string
            format: uuid
    RemoveTracksFromPlaylistRequest:
      type: object
      properties:
        track_ids:
          type: array
          items:
            type: string
            format: uuid
    ExportPlaylistRequest:
      type: object
      properties:
        format:
          type: string
          enum: [m3u, pls, xspf, json]
          default: m3u
        output_path:
          type: string
    ExportPlaylistResponse:
      type: object
      properties:
        file_path:
          type: string
        format:
          type: string
...